Make the addition of the courses link more flexible
We relied very specifically on a set of links to be present in the
user menu to add the "My courses" link. This would break if another
module tried to do the same. It also did break when a new link was
added by the distro.

This commit changes how the link is added by trying to find a
reference point and inserting the link before that. This will
properly add new links if other modules enable them. If the link
can not be found then we guess the position.

This guessing could be changed for a do while loop with multiple
guesses to make it more robust. Ideally the distribution would
allow us to define links with a section and a weight and add them
together itself.