Performance: cache material type alter result per request
Calling alter hooks is relatively expensive so you want to cache
the result. Especially when it's used in functions that's called
more than 100 times per request.

To achieve this we make a separate helper function that retrieves
the material types and stores them in a drupal static variable.
This has the added benefit of deduplicating code in case we want
to add another default type.