Performance: Split up block access check by block type
We split the block access check by block type to achieve a few
things.

1. It reduces the number of if statements that are being checked
even when a match is already found.
2. It ensures that only code that is needed to check access for the
current block is ever run and no other code.
3. The new structure of the functions ensures that we can bail early
if at any point we figure out that we can't do the access check.
This has the added benefit of clarifying the conditions for a block
to be hidden.