Turn explicit access checks into URL render array
By using the render array of the URL as part of our own link we will
inherit the access checks for the URL automatically and don't need to do
them ourselves. This makes our own code simpler, more extendable (e.g.
if someone changes the access rules) and makes it easier to cache.