TB-1934 Clean up properly when sections/materials are deleted
This commit ensures that no stray enrollments for users are left when a
part is deleted as this causes the overview that counts these
enrollments to be incorrect. It also ensures that a section doesn't
reference a material that no longer exists as this can cause the user to
be redirected to a page that isn't there.