TB-1884 Replace Access Denied with a redirect (#5)
Instead of giving the user an error that leaves them stuck when they try
to start a course for the second time, simply redirect them to the first
material in the course. From there they can find their way around.

The most likely scenario where they visit the link twice will also be
when they are still at the first part.