Ensure a user can't receive an access denied page
When a material of a section is finished but the next section is
not set to “In Progress” yet then the continue link will go to the
next material directly but this will cause an access denied in the
CourseWrapper::materialAccess because there are no enrollments yet.

In order to avoid this we check whether there is already an enrollment
for the next material and base our decision on where to send the user
based on this. The status of the current material doesn't really matter
anyway.