Fix missing label for unreachable materials
The fix for failing link generation was to remove the link entirely if there was no URL. However, this causes the material label not to be shown at all. The URL will be absent if a user is not allow to visit the Material. This can occur when a user has not reached that material yet. In those cases the material should still be shown to the user as a future goal.